[HWKMETRICS-396] exclude stats fields when buckets are empty

+++ core/metrics-core-service/src/test/java/org/hawkular/metrics/core/service/MetricsServiceITest.java
@@ -867,7 +867,7 @@ public class MetricsServiceITest extends MetricsITest {
 
         NumericBucketPoint bucket = actual.get(0);
 
-        assertEquals(testSize, bucket.getSamples());
+        assertEquals((Integer) testSize, bucket.getSamples());
         assertEquals(percentiles.size(), bucket.getPercentiles().size());
         assertEquals(top.getResult(), bucket.getPercentiles().get(3).getValue());
     }
